chore: add simple test for verify_rewards
