+++ rewards/calc_rewards.py
@@ -224,7 +224,7 @@ def generate_rewards_in_range(
 
     file_name = f"rewards-{chain_id}-{encode_hex(root_hash)}.json"
 
-    verify_rewards(past_tree, merkle_tree, tree_manager, chain)
+    verify_rewards(past_tree, merkle_tree, chain)
     if save:
         with open(file_name, "w") as fp:
             json.dump(merkle_tree, fp, indent=4)
