--- tests/test_utils/test_token_utils.py
@@ -1,16 +1,0 @@
-from rewards.rewards_checker import token_diff_table_item
-
-
-def test_token_diff_table():
-    table_item = token_diff_table_item(
-        name="bBADGER",
-        before=100,
-        after=200,
-        decimals=1
-    )
-    assert table_item == (100, [
-        'bBADGER',
-        '10.000000000000000000',
-        '20.000000000000000000',
-        '10.000000000000000000'
-    ])
